requirements.txt
numpy>=1.24
xarray>=2023.1
zarr>=2.14
gcsfs>=2023.1
scikit-learn>=1.2
scipy>=1.10
matplotlib>=3.7
pyyaml>=6.0
pytest>=7.0
